+++ ChestsAnywhere/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## Upcoming release
 * Farmhands in multiplayer can now access chests in all synced locations (generally their current location, the farm, the main player's farmhouse, and constructed farm buildings).
+* The menu now defaults to a chest in the current location, if any.
 * Updated Android support (thanks to ZaneYork!).
 * Improved translations. Thanks to D0n-A (updated Russian) and niniack (updated Chinese)!